+++ src/neo4j_graphrag/utils/validation.py
@@ -14,7 +14,8 @@
 #  limitations under the License.
 from __future__ import annotations
 
-from typing import Optional
+import importlib
+from typing import Optional, Tuple, Union, cast, Type
 
 
 def validate_search_query_input(
@@ -22,3 +23,23 @@ def validate_search_query_input(
 ) -> None:
     if not (bool(query_vector) ^ bool(query_text)):
         raise ValueError("You must provide exactly one of query_vector or query_text.")
+
+
+def issubclass_safe(
+    cls: Type[object], class_or_tuple: Union[Type[object], Tuple[Type[object]]]
+) -> bool:
+    if isinstance(class_or_tuple, tuple):
+        return any(issubclass_safe(cls, base) for base in class_or_tuple)
+
+    if issubclass(cls, class_or_tuple):
+        return True
+
+    # Handle case where module was reloaded
+    cls_module = importlib.import_module(cls.__module__)
+    # Get the latest version of the base class from the module
+    latest_base = getattr(cls_module, class_or_tuple.__name__, None)
+    latest_base = cast(Union[tuple[Type[object], ...], Type[object]], latest_base)
+    if issubclass(cls, latest_base):
+        return True
+
+    return False